requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
feedparser==6.0.8
Flask==2.0.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==2.4.0
google-auth==2.3.3
google-cloud-dialogflow==2.11.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
sgmllib3k==1.0.0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.2
